Move LogEntry from ExtVMFace.h to libethcore
