+++ libethcore/Common.h
@@ -31,6 +31,10 @@
 
 namespace dev
 {
+
+class RLP;
+class RLPStream;
+
 namespace eth
 {
 
@@ -96,6 +100,74 @@ enum class BlockPolarity
 	Dead,
 	Live
 };
+
+struct LogEntry
+{
+	LogEntry() = default;
+	explicit LogEntry(RLP const& _r);
+	LogEntry(Address const& _address, h256s _ts, bytes&& _d): address(_address), topics(std::move(_ts)), data(std::move(_d)) {}
+
+	void streamRLP(RLPStream& _s) const;
+
+	LogBloom bloom() const;
+
+	Address address;
+	h256s topics;
+	bytes data;
+};
+
+using LogEntries = std::vector<LogEntry>;
+
+struct LocalisedLogEntry: public LogEntry
+{
+	LocalisedLogEntry() = default;
+	explicit LocalisedLogEntry(LogEntry const& _le): LogEntry(_le) {}
+
+	LocalisedLogEntry(LogEntry const& _le, h256 _special):
+		LogEntry(_le),
+		isSpecial(true),
+		special(_special)
+	{}
+
+	LocalisedLogEntry(
+		LogEntry const& _le,
+		h256 const& _blockHash,
+		BlockNumber _blockNumber,
+		h256 const& _transactionHash,
+		unsigned _transactionIndex,
+		unsigned _logIndex,
+		BlockPolarity _polarity = BlockPolarity::Unknown
+	):
+		LogEntry(_le),
+		blockHash(_blockHash),
+		blockNumber(_blockNumber),
+		transactionHash(_transactionHash),
+		transactionIndex(_transactionIndex),
+		logIndex(_logIndex),
+		polarity(_polarity),
+		mined(true)
+	{}
+
+	h256 blockHash;
+	BlockNumber blockNumber = 0;
+	h256 transactionHash;
+	unsigned transactionIndex = 0;
+	unsigned logIndex = 0;
+	BlockPolarity polarity = BlockPolarity::Unknown;
+	bool mined = false;
+	bool isSpecial = false;
+	h256 special;
+};
+
+using LocalisedLogEntries = std::vector<LocalisedLogEntry>;
+
+inline LogBloom bloom(LogEntries const& _logs)
+{
+	LogBloom ret;
+	for (auto const& l: _logs)
+		ret |= l.bloom();
+	return ret;
+}
 
 class Transaction;
