+++ libevm/ExtVMFace.h
@@ -27,7 +27,6 @@
 #include <evm.h>
 #include <libdevcore/Common.h>
 #include <libdevcore/CommonData.h>
-#include <libdevcore/RLP.h>
 #include <libdevcore/SHA3.h>
 #include <libethcore/Common.h>
 #include <libethcore/BlockHeader.h>
@@ -85,81 +84,6 @@ public:
 private:
 	bytes m_bytes;
 };
-
-struct LogEntry
-{
-	LogEntry() = default;
-	explicit LogEntry(RLP const& _r) { address = (Address)_r[0]; topics = _r[1].toVector<h256>(); data = _r[2].toBytes(); }
-	LogEntry(Address const& _address, h256s _ts, bytes&& _d): address(_address), topics(std::move(_ts)), data(std::move(_d)) {}
-
-	void streamRLP(RLPStream& _s) const { _s.appendList(3) << address << topics << data; }
-
-	LogBloom bloom() const
-	{
-		LogBloom ret;
-		ret.shiftBloom<3>(sha3(address.ref()));
-		for (auto t: topics)
-			ret.shiftBloom<3>(sha3(t.ref()));
-		return ret;
-	}
-
-	Address address;
-	h256s topics;
-	bytes data;
-};
-
-using LogEntries = std::vector<LogEntry>;
-
-struct LocalisedLogEntry: public LogEntry
-{
-	LocalisedLogEntry() = default;
-	explicit LocalisedLogEntry(LogEntry const& _le): LogEntry(_le) {}
-
-	LocalisedLogEntry(LogEntry const& _le, h256 _special):
-		LogEntry(_le),
-		isSpecial(true),
-		special(_special)
-	{}
-
-	LocalisedLogEntry(
-		LogEntry const& _le,
-		h256 const& _blockHash,
-		BlockNumber _blockNumber,
-		h256 const& _transactionHash,
-		unsigned _transactionIndex,
-		unsigned _logIndex,
-		BlockPolarity _polarity = BlockPolarity::Unknown
-	):
-		LogEntry(_le),
-		blockHash(_blockHash),
-		blockNumber(_blockNumber),
-		transactionHash(_transactionHash),
-		transactionIndex(_transactionIndex),
-		logIndex(_logIndex),
-		polarity(_polarity),
-		mined(true)
-	{}
-
-	h256 blockHash;
-	BlockNumber blockNumber = 0;
-	h256 transactionHash;
-	unsigned transactionIndex = 0;
-	unsigned logIndex = 0;
-	BlockPolarity polarity = BlockPolarity::Unknown;
-	bool mined = false;
-	bool isSpecial = false;
-	h256 special;
-};
-
-using LocalisedLogEntries = std::vector<LocalisedLogEntry>;
-
-inline LogBloom bloom(LogEntries const& _logs)
-{
-	LogBloom ret;
-	for (auto const& l: _logs)
-		ret |= l.bloom();
-	return ret;
-}
 
 struct SubState
 {
